chore: rename ovis.ts to ast-to-query.ts
